Feature/neighborhoods lin (#187)
* fixed include

* Count constraint

Count neighborhood

* clang-format

* typo

* intLinLe neighborhood

* IntLinLe

* fixing tests

* fixing tests

* fixing tests

* fixing gcc bug

* checking same coeff

using SCD if all coeffs are identical for LinEQ. This should probably be done for all Linear constraints

+++ src/propagation/invariants/count.cpp
@@ -79,8 +79,7 @@ void Count::close(Timestamp ts) {
   lb = std::max(lb, _solver.lowerBound(_needle));
   ub = std::max(ub, _solver.lowerBound(_needle));
 
-  _counts.resize(static_cast<unsigned long>(ub - lb + 1),
-                 CommittableInt(ts, 0));
+  _counts.resize(static_cast<size_t>(ub - lb + 1), CommittableInt(ts, 0));
   _offset = lb;
 }
